Add more validation tests
