+++ jest.config.js
@@ -16,13 +16,15 @@
 
 module.exports = {
   collectCoverageFrom: [
-    // Collect coverage for all typescript files
+    // Collect coverage for all typescript files.
     "**/*.{ts,tsx}",
     // Ignore files in the src directory. Mainly, `index.ts` which should only
     // be exports.
     "!**/src/*",
     // Ignore any typescript declaration files.
     "!**/*.d.ts",
+    // ignore tests and snapshot tests.
+    "!**/*.test.{ts,tsx}",
   ],
   coverageReporters: ["lcov", "text"],
   projects: ["<rootDir>/packages/*"],
