+++ packages/pipeline-services/src/validation/index.test.ts
@@ -14,332 +14,9 @@
  * limitations under the License.
  */
 
-import { checkCircularReferences } from "./";
+import { validate } from "./";
 
-const linkExamples = [
-  // ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │
-  // ╰───╯      ╰───╯
-  //   ▲          │
-  //   ╰────l1────╯
-  {
-    it: "should detect a simple cycle",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "a", type: "nodeLink" },
-    ],
-    expected: ["l0", "l1"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮
-  // │ f │──c0─►│ a │──l0─►│ b │
-  // ╰───╯      ╰───╯      ╰───╯
-  //              ▲          │
-  //             l4         l1
-  //              │          ▼
-  //            ╭───╮      ╭───╮      ╭───╮
-  //            │ e │◄─l3──│ c │──l2─►│ d │
-  //            ╰───╯      ╰───╯      ╰───╯
-  //              ▲
-  //             c1
-  //              │
-  //            ╭───╮
-  //            │ g │
-  //            ╰───╯
-  {
-    it: "should ignore comment links",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l2", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "c", trgNodeId: "e", type: "nodeLink" },
-      { id: "l4", srcNodeId: "e", trgNodeId: "a", type: "nodeLink" },
-      { id: "c0", srcNodeId: "f", trgNodeId: "a", type: "commentLink" },
-      { id: "c1", srcNodeId: "g", trgNodeId: "c", type: "commentLink" },
-    ],
-    expected: ["l0", "l1", "l3", "l4"],
-  },
-
-  //    ╭───╮       ╭───╮
-  //  ╭─│ a │◄╮   ╭─│ c │◄╮
-  //  │ ╰───╯ │   │ ╰───╯ │
-  //  │       │   │       │
-  // l0      l1  l2      l3
-  //  │       │   │       │
-  //  │ ╭───╮ │   │ ╭───╮ │
-  //  ╰►│ b │─╯   ╰►│ d │─╯
-  //    ╰───╯       ╰───╯
-  {
-    it: "should detect multiple simple cycles",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "a", type: "nodeLink" },
-      { id: "l2", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "d", trgNodeId: "c", type: "nodeLink" },
-    ],
-    expected: ["l0", "l1", "l2", "l3"],
-  },
-
-  //   ╭────l4────╮
-  //   ▼          │
-  // ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │    ╭─│ e │◄╮
-  // ╰───╯      ╰───╯    │ ╰───╯ │
-  //   │          ▲      │       │
-  //  l2         l1     l5      l6
-  //   ▼          │      │       │
-  // ╭───╮      ╭───╮    │ ╭───╮ │
-  // │ d │◄─l3──│ c │    ╰►│ f │─╯
-  // ╰───╯      ╰───╯      ╰───╯
-  {
-    it: "should detect multiple cycles in a complex graph",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "c", trgNodeId: "b", type: "nodeLink" },
-      { id: "l2", srcNodeId: "a", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l4", srcNodeId: "b", trgNodeId: "a", type: "nodeLink" },
-      { id: "l5", srcNodeId: "e", trgNodeId: "f", type: "nodeLink" },
-      { id: "l6", srcNodeId: "f", trgNodeId: "e", type: "nodeLink" },
-    ],
-    expected: ["l0", "l4", "l5", "l6"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │◄─l1──│ c │    ╭─│ f │◄╮
-  // ╰───╯      ╰───╯      ╰───╯    │ ╰───╯ │
-  //              │          │      │       │
-  //             l4         l3     l6      l7
-  //              ▼          ▼      │       │
-  //            ╭───╮      ╭───╮    │ ╭───╮ │
-  //            │ d │◄─l5──│ e │    ╰►│ g │─╯
-  //            ╰───╯      ╰───╯      ╰───╯
-  //              │          ▲
-  //              ╰────l2────╯
-  {
-    it: "should detect multiple cycles in a complex graph - shuffled",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "c", trgNodeId: "b", type: "nodeLink" },
-      { id: "l2", srcNodeId: "d", trgNodeId: "e", type: "nodeLink" },
-      { id: "l3", srcNodeId: "c", trgNodeId: "e", type: "nodeLink" },
-      { id: "l4", srcNodeId: "b", trgNodeId: "d", type: "nodeLink" },
-      { id: "l5", srcNodeId: "e", trgNodeId: "d", type: "nodeLink" },
-      { id: "l6", srcNodeId: "f", trgNodeId: "g", type: "nodeLink" },
-      { id: "l7", srcNodeId: "g", trgNodeId: "f", type: "nodeLink" },
-    ],
-    expected: ["l2", "l5", "l6", "l7"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │◄─l1──│ c │
-  // ╰───╯      ╰───╯      ╰───╯
-  //              │          ▲
-  //              ╰────l2────╯
-  {
-    it: "should only detect links contributing to the cycle",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "c", trgNodeId: "b", type: "nodeLink" },
-      { id: "l2", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-    ],
-    expected: ["l1", "l2"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │──l2─►│ c │
-  // ╰───╯      ╰───╯      ╰───╯
-  //   ▲         │ ▲         │
-  //   ╰────l1───╯ ╰───l3────╯
-  {
-    it: "should detect joined cycles",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "a", type: "nodeLink" },
-      { id: "l2", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l3", srcNodeId: "c", trgNodeId: "b", type: "nodeLink" },
-    ],
-    expected: ["l0", "l1", "l2", "l3"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │──l2─►│ c │
-  // ╰───╯      ╰───╯      ╰───╯
-  //   ▲         │ ▲         │
-  //   ╰────l1───╯ ╰───l3────╯
-  {
-    it: "should detect joined cycles - shuffled",
-    given: [
-      { id: "l3", srcNodeId: "c", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "a", type: "nodeLink" },
-      { id: "l2", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-    ],
-    expected: ["l0", "l1", "l2", "l3"],
-  },
-
-  // ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │
-  // ╰───╯      ╰───╯
-  //   │          │
-  //  l3         l1
-  //   ▼          ▼
-  // ╭───╮      ╭───╮
-  // │ d │◄─l2──│ c │
-  // ╰───╯      ╰───╯
-  {
-    it: "should not detect fake cycle",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l2", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "a", trgNodeId: "d", type: "nodeLink" },
-    ],
-    expected: [],
-  },
-
-  //   ╭────l4────╮
-  //   ▼          │
-  // ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │
-  // ╰───╯      ╰───╯
-  //   │          ▲
-  //  l2         l1
-  //   ▼          │
-  // ╭───╮      ╭───╮
-  // │ d │◄─l3──│ c │
-  // ╰───╯      ╰───╯
-  {
-    it: "should detect cycle when a fork that gets checked first is safe",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "c", trgNodeId: "b", type: "nodeLink" },
-      { id: "l2", srcNodeId: "a", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l4", srcNodeId: "b", trgNodeId: "a", type: "nodeLink" },
-    ],
-    expected: ["l0", "l4"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │──l1─►│ c │──l2─►│ d │──l3─►│ e │
-  // ╰───╯      ╰───╯      ╰───╯      ╰───╯      ╰───╯
-  //                         │
-  //                        l4
-  //                         ▼
-  //                       ╭───╮      ╭───╮      ╭───╮      ╭───╮      ╭───╮
-  //                       │ f │──l5─►│ g │──l6─►│ h │──l7─►│ i │──l8─►│ j │
-  //                       ╰───╯      ╰───╯      ╰───╯      ╰───╯      ╰───╯
-  //                                                          ▲          │
-  //                                                          ╰────l9────╯
-  {
-    it: "should handle long forks",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l2", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "d", trgNodeId: "e", type: "nodeLink" },
-      { id: "l4", srcNodeId: "c", trgNodeId: "f", type: "nodeLink" },
-      { id: "l5", srcNodeId: "f", trgNodeId: "g", type: "nodeLink" },
-      { id: "l6", srcNodeId: "g", trgNodeId: "h", type: "nodeLink" },
-      { id: "l7", srcNodeId: "h", trgNodeId: "i", type: "nodeLink" },
-      { id: "l8", srcNodeId: "i", trgNodeId: "j", type: "nodeLink" },
-      { id: "l9", srcNodeId: "j", trgNodeId: "i", type: "nodeLink" },
-    ],
-    expected: ["l8", "l9"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │──l1─►│ c │──l8─►│ i │──l9─►│ j │
-  // ╰───╯      ╰───╯      ╰───╯      ╰───╯      ╰───╯
-  //                         │
-  //                        l2
-  //                         ▼
-  //                       ╭───╮      ╭───╮      ╭───╮      ╭───╮      ╭───╮
-  //                       │ d │──l3─►│ e │──l4─►│ f │──l5─►│ g │──l6─►│ h │
-  //                       ╰───╯      ╰───╯      ╰───╯      ╰───╯      ╰───╯
-  //                                                          ▲          │
-  //                                                          ╰────l7────╯
-  {
-    it: "should handle long forks - shuffled",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l2", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "d", trgNodeId: "e", type: "nodeLink" },
-      { id: "l4", srcNodeId: "e", trgNodeId: "f", type: "nodeLink" },
-      { id: "l5", srcNodeId: "f", trgNodeId: "g", type: "nodeLink" },
-      { id: "l6", srcNodeId: "g", trgNodeId: "h", type: "nodeLink" },
-      { id: "l7", srcNodeId: "h", trgNodeId: "g", type: "nodeLink" },
-      { id: "l8", srcNodeId: "c", trgNodeId: "i", type: "nodeLink" },
-      { id: "l9", srcNodeId: "i", trgNodeId: "j", type: "nodeLink" },
-    ],
-    expected: ["l6", "l7"],
-  },
-
-  // ╭───╮      ╭───╮      ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │──l1─►│ c │──l5─►│ f │──l6─►│ g │
-  // ╰───╯      ╰───╯      ╰───╯      ╰───╯      ╰───╯
-  //                         │          ▲
-  //                        l2         l4
-  //                         ▼          │
-  //                       ╭───╮      ╭───╮
-  //                       │ d │──l3─►│ e │
-  //                       ╰───╯      ╰───╯
-  {
-    it: "should not have a bug that a previous implementation had",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l2", srcNodeId: "c", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "d", trgNodeId: "e", type: "nodeLink" },
-      { id: "l4", srcNodeId: "e", trgNodeId: "f", type: "nodeLink" },
-      { id: "l5", srcNodeId: "c", trgNodeId: "f", type: "nodeLink" },
-      { id: "l6", srcNodeId: "f", trgNodeId: "g", type: "nodeLink" },
-    ],
-    expected: [],
-  },
-
-  //            ╭───╮
-  //            │ e │
-  //            ╰───╯
-  //              ▲
-  //             l3
-  //              │
-  // ╭───╮      ╭───╮      ╭───╮
-  // │ a │──l0─►│ b │──l2─►│ d │
-  // ╰───╯      ╰───╯      ╰───╯
-  //              │          │
-  //             l1          l4
-  //              ▼          ▼
-  //            ╭───╮      ╭───╮      ╭───╮
-  //            │ c │──l5─►│ f │──l6─►│ g │
-  //            ╰───╯      ╰───╯      ╰───╯
-  {
-    it: "issue #1303",
-    given: [
-      { id: "l0", srcNodeId: "a", trgNodeId: "b", type: "nodeLink" },
-      { id: "l1", srcNodeId: "b", trgNodeId: "c", type: "nodeLink" },
-      { id: "l2", srcNodeId: "b", trgNodeId: "d", type: "nodeLink" },
-      { id: "l3", srcNodeId: "b", trgNodeId: "e", type: "nodeLink" },
-      { id: "l4", srcNodeId: "d", trgNodeId: "f", type: "nodeLink" },
-      { id: "l5", srcNodeId: "c", trgNodeId: "f", type: "nodeLink" },
-      { id: "l6", srcNodeId: "f", trgNodeId: "g", type: "nodeLink" },
-    ],
-    expected: [],
-  },
-];
-
-describe("checkCircularReferences", () => {
-  for (const { it: should, given, expected } of linkExamples) {
-    // eslint-disable-next-line jest/valid-title
-    it(should, () => {
-      const startTime = Date.now();
-      const actual = checkCircularReferences(given);
-      const endTime = Date.now();
-
-      expect(new Set(actual)).toEqual(new Set(expected));
-      expect(endTime - startTime).toBeLessThan(300);
-    });
-  }
+it("should return an empty array for nothing", () => {
+  const problems = validate("", {});
+  expect(problems).toEqual([]);
 });
