+++ packages/pipeline-services/src/validation/index.ts
@@ -16,178 +16,86 @@
 
 import { parseTree, findNodeAtLocation } from "jsonc-parser";
 
-// 3 seconds should be ample time.
-// should only protect against any infinite loops.
-const TIMEOUT = 3000;
+import checkCircularReferences from "./check-circular-references";
+import { PartialProblem, Problem } from "./types";
+import { findNode, getLinks, getNodes } from "./utils";
 
-export interface Link {
-  id: string;
-  trgNodeId: string;
-  srcNodeId: string;
-  type: string;
-  path?: any[];
-}
+export function getLinkProblems(pipeline: any) {
+  const links = getLinks(pipeline);
 
-export function checkCircularReferences(links: Link[]) {
-  const startTime = Date.now();
+  const taintedLinks = checkCircularReferences(links);
 
-  // filter out comment links.
-  links = links.filter((l) => l.type !== "commentLink");
-
-  // organize links into easily indexable map:
-  // {srcNodeId: link[]}
-  const linkMap: { [key: string]: Link[] } = {};
-  for (const l of links) {
-    if (linkMap[l.srcNodeId] === undefined) {
-      linkMap[l.srcNodeId] = [];
-    }
-    linkMap[l.srcNodeId].push(l);
-  }
-
-  let orderedChain: string[] = [];
-  const taintedLinks = new Set<string>();
-  const seen = new Set<string>();
-  const stack: Link[] = [];
-  for (const l of links) {
-    if (seen.has(l.id)) {
+  let problems: PartialProblem[] = [];
+  for (const linkID of taintedLinks) {
+    const link = links.find((l) => l.id === linkID);
+    if (link === undefined) {
       continue;
     }
-    seen.add(l.id);
-    orderedChain = [];
-    orderedChain.push(l.id);
-    const linksToVisit = linkMap[l.trgNodeId];
-    if (linksToVisit === undefined) {
+
+    const source = findNode(pipeline, link.srcNodeId);
+    const target = findNode(pipeline, link.trgNodeId);
+    problems.push({
+      message: `The connection between nodes '${source.app_data.ui_data.label}' and '${target.app_data.ui_data.label}' is part of a circular reference.`,
+      path: [...link.path, "node_id_ref"],
+      info: {
+        type: "circularReference",
+        pipelineID: pipeline.id,
+        linkID: linkID,
+      },
+    });
+  }
+
+  return problems;
+}
+
+export function getNodeProblems(pipeline: any, nodeDefinitions: any) {
+  const nodes = getNodes(pipeline);
+  let problems: PartialProblem[] = [];
+  for (const [n, node] of nodes.entries()) {
+    if (node.type !== "execution_node") {
       continue;
     }
-    stack.push(...linksToVisit);
-    const forkStack: number[] = [];
-    const chainLength = orderedChain.length;
-    forkStack.push(...linksToVisit.map(() => chainLength));
 
-    while (stack.length > 0 && Date.now() - startTime < TIMEOUT) {
-      forkStack.pop();
-      const l = stack.pop()!; // we should be gauranteed an item.
-      seen.add(l.id);
-      const seenLinkIndex = orderedChain.indexOf(l.id);
+    const nodeDef = nodeDefinitions.find((n: any) => n.op === node.op);
+    if (nodeDef === undefined) {
+      continue;
+    }
 
-      // We hit a link we've already seen in the chain. This means there is a
-      // cycle from the seen link to the end of the chain.
-      if (seenLinkIndex > -1) {
-        for (const item of orderedChain.slice(seenLinkIndex)) {
-          taintedLinks.add(item);
-        }
-
-        // This is completely useless, but it makes me feel better that this is
-        // here.
-        const position = forkStack.slice(-1).pop();
-        if (position !== undefined) {
-          orderedChain = orderedChain.slice(0, position);
-        }
-        // if position is undefined we will set orderedChain to []
-        continue;
+    for (const prop of nodeDef.properties?.parameters ?? []) {
+      // If the property isn't in the json, report the error one level higher.
+      let path = ["nodes", n, "app_data"];
+      if (node.app_data[prop.id] !== undefined) {
+        path.push(prop.id);
       }
 
-      orderedChain.push(l.id);
-
-      const linksToVisit = linkMap[l.trgNodeId];
-
-      // We reached the end of a chain.
-      if (linksToVisit === undefined) {
-        // This is 100% necessary unlike the other example.
-        const position = forkStack.slice(-1).pop();
-        if (position !== undefined) {
-          orderedChain = orderedChain.slice(0, position);
+      // this should be safe because a boolean can't be required
+      // otherwise we would need to check strings for undefined or empty string
+      // NOTE: 0 is also falsy, but we don't have any number inputs right now?
+      // TODO: We should update this to do type checking.
+      if (prop.required && !node.app_data[prop.id]) {
+        const paramInfo = nodeDef.properties?.uihints.parameter_info;
+        const param = paramInfo.find((p: any) => p.parameter_ref === prop.id);
+        if (param === undefined) {
+          continue;
         }
-        // if position is undefined we will set orderedChain to []
-        continue;
+        problems.push({
+          message: `The property '${param.label.default}' on node '${node.app_data.ui_data.label}' is required.`,
+          path,
+          info: {
+            type: "missingProperty",
+            pipelineID: pipeline.id,
+            nodeID: node.id,
+            property: prop.id,
+          },
+        });
       }
-
-      // Uncharted teritory, add it to the stack to be explored.
-      stack.push(...linksToVisit);
-      const chainLength = orderedChain.length;
-      forkStack.push(...linksToVisit.map(() => chainLength));
     }
   }
 
-  return [...taintedLinks];
+  return problems;
 }
 
-enum Severity {
-  Error = 1,
-  Warning,
-  Info,
-  Hint,
-}
-
-interface CircularReferenceInfo {
-  type: "circularReference";
-  pipelineID: string;
-  linkID: string;
-}
-
-interface MissingPropertyInfo {
-  type: "missingProperty";
-  pipelineID: string;
-  nodeID: string;
-  property: string;
-}
-
-interface InvalidNodeInfo {
-  type: "invalidNode";
-  pipelineID: string;
-  nodeID: string;
-  op: string;
-}
-
-interface Problem {
-  severity: Severity;
-  range: {
-    offset: number;
-    length: number;
-  };
-  message: string;
-  info: CircularReferenceInfo | MissingPropertyInfo | InvalidNodeInfo;
-}
-
-function isNode(nodes: any[], id: string) {
-  const index = nodes.findIndex((node) => node.id === id);
-  if (index === -1) {
-    return false;
-  }
-  return true;
-}
-
-function getLinks(pipeline: any): Link[] {
-  let links: Link[] = [];
-  for (const [n, node] of pipeline.nodes.entries()) {
-    for (const [i, input] of node.inputs?.entries() ?? []) {
-      if (input.links !== undefined) {
-        for (const [l, link] of input.links.entries()) {
-          if (isNode(pipeline.nodes, link.node_id_ref)) {
-            links.push({
-              id: link.id,
-              trgNodeId: node.id,
-              srcNodeId: link.node_id_ref,
-              type: "",
-              path: ["nodes", n, "inputs", i, "links", l],
-            });
-          }
-        }
-      }
-    }
-  }
-  return links;
-}
-
-function getNodes(pipeline: any): any[] {
-  return pipeline.nodes;
-}
-
-function findNode(pipeline: any, id: string) {
-  return pipeline.nodes.find((node: any) => node.id === id);
-}
-
-export function validate(pipeline: string, nodeDefinitions: any): Problem[] {
+export function validate(pipeline: string, nodeDefinitions: any) {
   const pipelineTreeRoot = parseTree(pipeline);
   if (pipelineTreeRoot === undefined) {
     return [];
@@ -197,77 +105,28 @@ export function validate(pipeline: string, nodeDefinitions: any): Problem[] {
 
   let problems: Problem[] = [];
   for (const [p, pipeline] of pipelineJSON.pipelines.entries()) {
-    const links = getLinks(pipeline);
-    const taintedLinks = checkCircularReferences(links);
-    const linkProblems = taintedLinks.map(
-      (linkID): Problem => {
-        const link = links.find((l) => l.id === linkID);
+    let partials: PartialProblem[] = [];
 
-        const path = ["pipelines", p, ...(link?.path ?? []), "node_id_ref"];
-        const location = findNodeAtLocation(pipelineTreeRoot, path);
-        const source = findNode(pipeline, link?.srcNodeId ?? "");
-        const target = findNode(pipeline, link?.trgNodeId ?? "");
+    partials.push(...getLinkProblems(pipeline));
+    partials.push(...getNodeProblems(pipeline, nodeDefinitions));
+
+    const getLocation = (path: any[]) => {
+      return findNodeAtLocation(pipelineTreeRoot, ["pipelines", p, ...path]);
+    };
+
+    problems.push(
+      ...partials.map((p) => {
+        const { path, ...rest } = p;
         return {
+          ...rest,
           severity: 1,
-          message: `The connection between nodes '${source.app_data.ui_data.label}' and '${target.app_data.ui_data.label}' is part of a circular reference.`,
           range: {
-            offset: location?.offset ?? 0,
-            length: location?.length ?? 0,
-          },
-          info: {
-            type: "circularReference",
-            pipelineID: pipeline.id,
-            linkID: linkID,
+            offset: getLocation(path)?.offset ?? 0,
+            length: getLocation(path)?.length ?? 0,
           },
         };
-      }
+      })
     );
-    problems.push(...linkProblems);
-
-    // validate properties
-    const nodes = getNodes(pipeline);
-    for (const [n, node] of nodes.entries()) {
-      if (node.op === undefined) {
-        // NOTE: supernode or binding, don't know if we need to validate anything?
-        continue;
-      }
-      const nodeDef = nodeDefinitions.find((n: any) => n.op === node.op);
-      if (nodeDef) {
-        for (const prop of nodeDef.properties?.parameters ?? []) {
-          // If the property isn't in the json, report the error one level higher.
-          let path = ["pipelines", p, "nodes", n, "app_data"];
-          if (node.app_data[prop.id] !== undefined) {
-            path.push(prop.id);
-          }
-          const location = findNodeAtLocation(pipelineTreeRoot, path);
-
-          // this should be safe because a boolean can't be required
-          // otherwise we would need to check strings for undefined or empty string
-          // NOTE: 0 is also falsy, but we don't have any number inputs right now?
-          // TODO: We should update this to do type checking.
-          if (prop.required && !node.app_data[prop.id]) {
-            const label = nodeDef.properties?.uihints.parameter_info.find(
-              (p: any) => p.parameter_ref === prop.id
-            )?.label.default;
-            problems.push({
-              severity: 1,
-              message: `The property '${label}' on node '${node.app_data.ui_data.label}' is required.`,
-              range: {
-                offset: location?.offset ?? 0,
-                length: location?.length ?? 0,
-              },
-              info: {
-                type: "missingProperty",
-                pipelineID: pipeline.id,
-                nodeID: node.id,
-                property: prop.id,
-              },
-            });
-            continue;
-          }
-        }
-      }
-    }
   }
 
   return problems;
